requirements.txt
av==12.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorzero==2.0
contourpy==1.2.1
cycler==0.12.1
evdev==1.7.1
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.6.0
gpiozero==2.0.1
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
lgpio==0.2.2.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pandas==2.2.2
picamera2==0.3.18
pidng==4.0.9
piexif==1.1.3
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
pynput==1.7.7
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-prctl==1.8.1
python-xlib==0.33
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rpi-lgpio==0.6
rpi-libcamera==0.1a4
scipy==1.13.1
seaborn==0.13.2
simplejpeg==1.7.2
six==1.16.0
sympy==1.12.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.30
ultralytics-thop==0.2.8
urllib3==2.2.1
v4l2-python3==0.3.4
